requirements.txt
# HN Daily Summary Dependencies
openai>=1.0.0      # DeepSeek API is OpenAI-compatible
requests>=2.31.0   # HTTP requests for HN API
python-dotenv>=1.0.0  # Load environment variables
httpx[socks]       # SOCKS proxy support (optional, for users behind proxy)
feedparser>=6.0.0  # RSS/Atom feed parsing
pyyaml>=6.0.0      # YAML config parsing
